+++ plugins/modules/win_combine_certificates.ps1
@@ -3,18 +3,18 @@
 #AnsibleRequires -CSharpUtil Ansible.Basic
 
 $spec = @{
-  options = @{
-    certificates = @{
-      type = 'list'
-      elements = 'str'
-      required = $true
+    options = @{
+        certificates = @{
+            type = 'list'
+            elements = 'path'
+            required = $true
+        }
+        path = @{
+            type = 'path'
+            required = $true
+        }
     }
-    path = @{
-      type = 'str'
-      required = $true
-    }
-  }
-  supports_check_mode = $true
+    supports_check_mode = $true
 }
 
 $module = [Ansible.Basic.AnsibleModule]::Create($args, $spec)
@@ -22,44 +22,38 @@ $module = [Ansible.Basic.AnsibleModule]::Create($args, $spec)
 $expectedContent = ''
 
 foreach ($certificate in $module.Params['certificates']) {
-  
-  if (-not ($expectedContent -eq '')) {
 
-    $expectedContent += "`n"
-  }
+    if (-not ($expectedContent -eq '')) {
+        $expectedContent += "`n"
+    }
 
-  if (-not (Test-Path -LiteralPath $certificate -PathType Leaf)) {
+    if (-not (Test-Path -LiteralPath $certificate -PathType Leaf)) {
+        $module.FailJson("Certificate file '$certificate' does not exist.")
+    }
 
-    $module.FailJson("Certificate file '$certificate' does not exist.")
-  }
+    $content = [System.IO.File]::ReadAllText($certificate)
 
-  $content = [System.IO.File]::ReadAllText($certificate)
-
-  $expectedContent += $content
+    $expectedContent += $content
 }
 
 if (Test-Path -LiteralPath $module.Params['path'] -PathType Container) {
-
-  $module.FailJson("Path '{$($module.Params['path'])}' is a directory.")
+    $module.FailJson("Path '{$($module.Params['path'])}' is a directory.")
 }
 
 if (-not (Test-Path -LiteralPath $module.Params['path'])) {
-
-  $actualContent = $null
+    $actualContent = $null
 }
 else {
-
-  $actualContent = [System.IO.File]::ReadAllText($module.Params['path'])
+    $actualContent = [System.IO.File]::ReadAllText($module.Params['path'])
 }
 
 if ($actualContent -ne $expectedContent) {
 
-  $module.Result["changed"] = $true
+    $module.Result["changed"] = $true
 
-  if (-not $module.CheckMode) {
-
-    [System.IO.File]::WriteAllText($module.Params['path'], $expectedContent)
-  }
+    if (-not $module.CheckMode) {
+        [System.IO.File]::WriteAllText($module.Params['path'], $expectedContent)
+    }
 }
 
 $module.ExitJson()