+++ plugins/modules/win_package_provider.ps1
@@ -23,6 +23,7 @@ $changed = $false
 $name = $module.Params['name']
 
 if ($module.Params.ContainsKey('min_version')) {
+
     try {
         $minVersion = [System.Version]::Parse($module.Params['min_version'])
     }
@@ -41,17 +42,16 @@ if ($providers.Count -gt 0) {
     $currentVersion = $providers[0].Version
 }
 
-if ($null -eq $currentVersion)
-{
+if ($null -eq $currentVersion) {
     $changed = $true
 }
 
-if ($null -ne $minVersion -and $currentVersion -lt $minVersion)
-{
+if ($null -ne $minVersion -and $currentVersion -lt $minVersion) {
     $changed = $true
 }
 
 if ($changed) {
+
     if ($null -eq $currentVersion) {
         $module.Result["previous_version"] = "Not Installed"
     }
@@ -60,15 +60,15 @@ if ($changed) {
     }
 
     if (-not $module.CheckMode) {
+
         $providers = ([Array](Install-PackageProvider -Name $name -Force -ForceBootstrap))
 
         $installedVersion = $providers[0].Version
-  
-        if ($null -ne $minVersion -and $installedVersion -lt $minVersion)
-        {
+
+        if ($null -ne $minVersion -and $installedVersion -lt $minVersion) {
             $module.FailJson("Failed to install package provider '$name' version '$minVersion'.")
         }
-  
+
         $module.Result["installed_version"] = $installedVersion.ToString()
     }
 }